Support more 10_10_10_2 formats as vertex buffers
Vertex input buffers only support VK_FORMAT_A2B10G10R10_UNORM_PACK32.
This cl adds all the signed and integer variants of that format.

This cl also removes StreamType and replaces it with VkFormat.

Tests: dEQP-VK.*r10*
Bug: b/142661203

+++ src/Pipeline/SamplerCore.hpp
@@ -100,15 +100,15 @@ private:
 
 	bool hasFloatTexture() const;
 	bool hasUnnormalizedIntegerTexture() const;
-	bool hasUnsignedTextureComponent(int component) const;
-	int textureComponentCount() const;
+	bool hasUnsignedTextureComponent(uint8_t component) const;
+	uint8_t textureComponentCount() const;
 	bool hasThirdCoordinate() const;
 	bool has16bitTextureFormat() const;
 	bool has8bitTextureComponents() const;
 	bool has16bitTextureComponents() const;
 	bool has32bitIntegerTextureComponents() const;
 	bool isYcbcrFormat() const;
-	bool isRGBComponent(int component) const;
+	bool isRGBComponent(uint8_t component) const;
 	bool borderModeActive() const;
 	bool isCube() const;
 	VkComponentSwizzle gatherSwizzle() const;
